Add the first implementation of line filtering
This is a custom superfence for pymdown-extensions that can filter lines
by specifying a set of line ranges to show.

The ranges are specified in the `show_lines` option, which is a
comma-separated list of ranges, where each range is in the format
`start:end` or `line`, where `start`, `end` and `line` are positive
integers.

If `start` is empty, it is assumed to be 1. If `end` is empty, it is
assumed to be the end of the file. If only `line` is used, it is assumed
to be both `start` and `end` (so allowing just one line).

Lines are 1-indexed.

Warnings are emitted if the `show_lines` option is invalid. There is a
hacky way to determine if we are running inside MkDocs or not, and if we
are then we use the MkDocs logger, so warnings are shown in the MkDocs
output with special formatting and also are detected as such when
running in *strict* mode. If we are not running inside MkDocs, then we
use the logger for this module.

Signed-off-by: Leandro Lucarella <luca-frequenz@llucax.com>