Fix mocking of `md.preprocessors` in tests
The `md.preprocessors` attribute wasn't being mocked correctly and
returned a new mock object instead of the custom mock object we wanted.

The tests passed anyway, because the only thing that was used was the
`highlight` return value, which evaluates to `True` when it is a new
mock object instead of a proper `bool` value, but now that the type of
the preprocessor is checked to be able to call `highlight`, the tests
failed.

Signed-off-by: Leandro Lucarella <luca-frequenz@llucax.com>